Fixes issue with passing the movieid to the viewmodel
