+++ composeApp/src/commonMain/kotlin/me/androidbox/busbymoviesv2/movie_details/di/movieDetailModule.kt
@@ -1,6 +1,7 @@
 package me.androidbox.busbymoviesv2.movie_details.di
 
 import io.ktor.client.HttpClient
+import me.androidbox.busbymoviesv2.configuration.domain.usecases.ConfigurationUseCase
 import me.androidbox.busbymoviesv2.movie_details.data.remote_data_source.MovieDetailRemoteDataSource
 import me.androidbox.busbymoviesv2.movie_details.data.remote_data_source.imp.MovieDetailRemoteDataSourceImp
 import me.androidbox.busbymoviesv2.movie_details.data.repository.MovieDetailRepository
@@ -8,7 +9,7 @@ import me.androidbox.busbymoviesv2.movie_details.data.repository.imp.MovieDetail
 import me.androidbox.busbymoviesv2.movie_details.domain.MovieDetailUseCase
 import me.androidbox.busbymoviesv2.movie_details.domain.imp.MovieDetailUseCaseImp
 import me.androidbox.busbymoviesv2.movie_details.presentation.MovieDetailViewModel
-import org.koin.compose.viewmodel.dsl.viewModelOf
+import org.koin.compose.viewmodel.dsl.viewModel
 import org.koin.dsl.module
 
 val movieDetailModule = module {
@@ -24,5 +25,13 @@ val movieDetailModule = module {
         MovieDetailRemoteDataSourceImp(get<HttpClient>())
     }
 
-    viewModelOf(::MovieDetailViewModel)
+    // viewModelOf(::MovieDetailViewModel)
+
+    viewModel { (movieId: Int) ->
+        MovieDetailViewModel(
+            movieDetailUseCase = get<MovieDetailUseCase>(),
+            configurationUseCase = get<ConfigurationUseCase>(),
+            movieId = movieId
+        )
+    }
 }
